requirements.txt
docopt == 0.6.2
python-dateutil == 2.8.1
python-json-logger == 0.1.11
requests == 2.24.0
unittest-xml-reporting == 3.0.2
